Struct glium_text::TextDisplay

pub struct TextDisplay<F> where F: Deref<Target=FontTexture> {
    // some fields omitted
}

Object that will allow you to draw a text.

Methods

impl<F> TextDisplay<F> where F: Deref<Target=FontTexture>

fn new(system: &TextSystem, texture: F, text: &str) -> TextDisplay<F>

Builds a new text display that allows you to draw text.

fn get_width(&self) -> f32

Returns the width in GL units of the text.

fn set_text(&mut self, text: &str)

Modifies the text on this display.