Struct glium_text::TextSystem

pub struct TextSystem {
    // some fields omitted
}

Object that contains the elements shared by all TextDisplay objects.

Required to create a TextDisplay.

Methods

impl TextSystem

fn new<F>(facade: &F) -> TextSystem where F: Facade

Builds a new text system that must be used to build TextDisplay objects.